新增DTO 目前只有把 子分類的 DTO  刪除 搜尋單筆 以父分類搜尋全部子分類

+++ src/main/java/com/outfit_share/service/product/SubcategoryService.java
@@ -1,12 +1,16 @@
 package com.outfit_share.service.product;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.hibernate.Hibernate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.outfit_share.entity.product.Subcategory;
+import com.outfit_share.entity.product.SubcategoryDTO;
 import com.outfit_share.repository.product.SubcategoryRepository;
 
 import jakarta.transaction.Transactional;
@@ -37,25 +41,45 @@ public class SubcategoryService {
 	}
 	
 //	刪除子分類
-	public void deleteSubcategory(Integer id) {
-		subcategoryRepository.deleteById(id);
+	public SubcategoryDTO deleteSubcategory(Integer id) {
+		Optional<Subcategory> optional = subcategoryRepository.findById(id);
+		
+		if (optional.isPresent()) {
+			
+			Subcategory subcategory = optional.get();
+			SubcategoryDTO subcategoryDTO = new SubcategoryDTO(subcategory);
+			subcategoryRepository.deleteById(id); 
+			return subcategoryDTO;
+		}
+		return null;
 	}
 	
 //	搜尋子分類
-	public Subcategory findSubcategoryById(Integer id) {
+	public SubcategoryDTO findSubcategoryById(Integer id) {
 		 Optional<Subcategory> optional = subcategoryRepository.findById(id);
 		 
+		 //更簡潔的寫法
+//		 return optional.map(SubcategoryDTO::new).orElse(null);
 		 if(optional.isPresent()) {
 			 
-			 return optional.get();
+			 Subcategory subcategory = optional.get();
+			 return new SubcategoryDTO(subcategory);
 		 }
-		 
 		 return null;
 	}
 	
-	// 新增: 通過父分類ID獲取子分類列表
-	public List<Subcategory> findSubcategoriesByCategoryId(Integer categoryId) {
-        return subcategoryRepository.findByCategoryCategoryId(categoryId);
-    }
+	
+	 // 新增: 通過父分類ID獲取子分類列表
+	 public List<SubcategoryDTO> findSubcategoriesByCategoryId(Integer categoryId) {
+	        List<Subcategory> list = subcategoryRepository.findByCategoryCategoryId(categoryId);
+	        List<SubcategoryDTO> dtolist = new ArrayList<>();
+	        
+	        for(Subcategory subcategory : list) {
+				Hibernate.initialize(subcategory.getCategory());
+				SubcategoryDTO subcategoryDTO = new SubcategoryDTO(subcategory);
+				dtolist.add(subcategoryDTO);
+	    }
+	        return dtolist;
+	 }
 	
 }
